Working on #92: rum_runner stop and kill

+++ lib/RUM/Platform/Local.pm
@@ -2,6 +2,7 @@ package RUM::Platform::Local;
 
 use strict;
 use warnings;
+use autodie;
 
 use Carp;
 use Text::Wrap qw(wrap fill);
@@ -582,16 +583,18 @@ sub _reads {
 
 sub stop {
     my ($self) = @_;
+
+    if ( ! $self->is_running ) {
+        $self->alert(
+            "There doesn't seem to be a RUM job running in ",
+            $self->config->output_dir());
+        return;
+    }
     my $lock_file = $self->config->lock_file;
-    -e $lock_file or die
-        join(" ", 
-             "There doesn't seem to be a RUM pipeline running in",
-             $self->config->output_dir(), 
-             "(I can't open the lock file)");
-    open my $in, "<", $lock_file or die 
-        "I can't read a pid from the lock file $lock_file";
+    open my $in, "<", $lock_file;
+
     my $pid = int(<$in>) or die 
-        "The lock file $lock_file exists but does not contain a pid";
+    "The lock file $lock_file exists but does not contain a pid";
     $self->say("Killing process $pid");
     kill 15, $pid or die "I can't kill $pid: $!";
 }
@@ -599,6 +602,12 @@ sub stop {
 sub job_report {
     my ($self) = @_;
     return RUM::JobReport->new($self->config);
+}
+
+sub is_running {
+    my ($self) = @_;
+    my $lock_file = $self->config->lock_file;
+    return -e $lock_file;
 }
 
 1;
